+++ core/chains/evm/config/v2/config.go
@@ -29,12 +29,17 @@ type HasEVMConfigs interface {
 
 type EVMConfigs []*EVMConfig
 
+func (cs EVMConfigs) DetectDuplicates(fs EVMConfigs) (err error) {
+	combined := append(cs, fs...)
+	return combined.ValidateConfig()
+}
+
 func (cs EVMConfigs) ValidateConfig() (err error) {
 	// Unique chain IDs
 	chainIDs := v2.UniqueStrings{}
 	for i, c := range cs {
 		if chainIDs.IsDupeFmt(c.ChainID) {
-			err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.ChainID", i), c.ChainID.String()))
+			err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("EVM.%d.ChainID", i), c.ChainID.String()))
 		}
 	}
 
@@ -43,7 +48,7 @@ func (cs EVMConfigs) ValidateConfig() (err error) {
 	for i, c := range cs {
 		for j, n := range c.Nodes {
 			if names.IsDupe(n.Name) {
-				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.Nodes.%d.Name", i, j), *n.Name))
+				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("EVM.%d.Nodes.%d.Name", i, j), *n.Name))
 			}
 		}
 	}
@@ -54,7 +59,7 @@ func (cs EVMConfigs) ValidateConfig() (err error) {
 		for j, n := range c.Nodes {
 			u := (*url.URL)(n.WSURL)
 			if wsURLs.IsDupeFmt(u) {
-				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.Nodes.%d.WSURL", i, j), u.String()))
+				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("EVM.%d.Nodes.%d.WSURL", i, j), u.String()))
 			}
 		}
 	}