+++ core/chains/solana/config.go
@@ -23,12 +23,17 @@ import (
 
 type SolanaConfigs []*SolanaConfig
 
+func (cs SolanaConfigs) DetectDuplicates(fs SolanaConfigs) (err error) {
+	combined := append(cs, fs...)
+	return combined.ValidateConfig()
+}
+
 func (cs SolanaConfigs) ValidateConfig() (err error) {
 	// Unique chain IDs
 	chainIDs := v2.UniqueStrings{}
 	for i, c := range cs {
 		if chainIDs.IsDupe(c.ChainID) {
-			err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.ChainID", i), *c.ChainID))
+			err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("Solana.%d.ChainID", i), *c.ChainID))
 		}
 	}
 
@@ -37,7 +42,7 @@ func (cs SolanaConfigs) ValidateConfig() (err error) {
 	for i, c := range cs {
 		for j, n := range c.Nodes {
 			if names.IsDupe(n.Name) {
-				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.Nodes.%d.Name", i, j), *n.Name))
+				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("Solana.%d.Nodes.%d.Name", i, j), *n.Name))
 			}
 		}
 	}
@@ -48,7 +53,7 @@ func (cs SolanaConfigs) ValidateConfig() (err error) {
 		for j, n := range c.Nodes {
 			u := (*url.URL)(n.URL)
 			if urls.IsDupeFmt(u) {
-				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.Nodes.%d.URL", i, j), u.String()))
+				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("Solana.%d.Nodes.%d.URL", i, j), u.String()))
 			}
 		}
 	}