+++ core/chains/starknet/config.go
@@ -19,12 +19,17 @@ import (
 
 type StarknetConfigs []*StarknetConfig
 
+func (cs StarknetConfigs) DetectDuplicates(fs StarknetConfigs) (err error) {
+	combined := append(cs, fs...)
+	return combined.ValidateConfig()
+}
+
 func (cs StarknetConfigs) ValidateConfig() (err error) {
 	// Unique chain IDs
 	chainIDs := v2.UniqueStrings{}
 	for i, c := range cs {
 		if chainIDs.IsDupe(c.ChainID) {
-			err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.ChainID", i), *c.ChainID))
+			err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("Starknet.%d.ChainID", i), *c.ChainID))
 		}
 	}
 
@@ -33,7 +38,7 @@ func (cs StarknetConfigs) ValidateConfig() (err error) {
 	for i, c := range cs {
 		for j, n := range c.Nodes {
 			if names.IsDupe(n.Name) {
-				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.Nodes.%d.Name", i, j), *n.Name))
+				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("Starknet.%d.Nodes.%d.Name", i, j), *n.Name))
 			}
 		}
 	}
@@ -44,7 +49,7 @@ func (cs StarknetConfigs) ValidateConfig() (err error) {
 		for j, n := range c.Nodes {
 			u := (*url.URL)(n.URL)
 			if urls.IsDupeFmt(u) {
-				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("%d.Nodes.%d.URL", i, j), u.String()))
+				err = multierr.Append(err, v2.NewErrDuplicate(fmt.Sprintf("Starknet.%d.Nodes.%d.URL", i, j), u.String()))
 			}
 		}
 	}